very small part of breakout game.
This was a fully realized game made by a team consisting of Derek Alexander, Ethan Gatlin, Steven Killen, Matthew Strawn and with plenty of help from our professor Scott Strentzsch aka our fearless leader. As this was a team project this is all I feel comfortable showing of our work which consist of over 60 files of code as well as art and sound assets. An important note, this was made our own hand written engine consisting of Windows and DirectX.